Django Project/requirements.txt
Django==2.2.20
django_bootstrap_datepicker_plus==3.0.5
django-bootstrap4==3.0.1
python_dateutil==2.8.2
